[#5217] Dissolve module into ProcessedSource & move back 4 methods
