+++ lib/rubocop.rb
@@ -75,7 +75,6 @@ require_relative 'rubocop/cop/force'
 require_relative 'rubocop/cop/severity'
 require_relative 'rubocop/cop/generator'
 require_relative 'rubocop/cop/generator/require_file_injector'
-require_relative 'rubocop/cop/processed_source_logic'
 
 require_relative 'rubocop/cop/variable_force'
 require_relative 'rubocop/cop/variable_force/branch'
