+++ lib/rubocop/cop/layout/empty_lines.rb
@@ -51,9 +51,7 @@ module RuboCop
               # we need to be wary of comments since they
               # don't show up in the tokens
               ((prev_line + 1)...cur_line).each do |line|
-                unless processed_source.previous_and_current_lines_empty?(line)
-                  next
-                end
+                next unless previous_and_current_lines_empty?(line)
 
                 yield source_range(processed_source.buffer, line, 0)
               end
@@ -65,6 +63,10 @@ module RuboCop
 
         def exceeds_line_offset?(line_diff)
           line_diff > LINE_OFFSET
+        end
+
+        def previous_and_current_lines_empty?(line)
+          processed_source[line - 2].empty? && processed_source[line - 1].empty?
         end
       end
     end
